pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "research_agent"
version = "0.1.0"
description = "A professional AI research agent using Gemini function calling"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]